requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14.0
Pillow>=9.0.0         # Xử lý ảnh (minh chứng học phí, ảnh khuôn mặt)
numpy>=1.22.0         # Dùng trong extract_face_vector
opencv-python>=4.6.0  # Dùng trong extract_face_vector
psycopg2-binary>=2.9.5 # Kết nối PostgreSQL
python-dotenv>=1.0.0  # Nếu có dùng file .env
django-cors-headers>=4.0.0  # Nếu dùng cho app mobile/web call API

# Nếu có dùng thêm các chức năng export excel:
openpyxl>=3.1.0
pandas>=2.0.0         # Nếu có dùng pandas cho xuất excel hoặc tính toán

# Nếu dùng JWT sau này:
djangorestframework-simplejwt>=5.2.0

# Nếu dùng face recognition bằng thư viện face_recognition:
face_recognition>=1.3.0
